'plugin' must be a real object with a get_networks attribute
Also needed to adapt code for a new change in Neutron master,
specifically the plugin passed to CalicoDeviceManager must now be a real
object with a get_networks attribute.  The value of that attribute is
passed to the interface driver, so that the interface driver can use it
if necessary.  In Calico's case, RoutedInterfaceDriver doesn't need to
call get_networks, so the attribute can be anything, so long as it is
defined.  Therefore the fix here is to add a get_networks attribute to
FakePlugin, and to construct CalicoDeviceManager with
plugin=FakePlugin().

Change-Id: I1db6e9e3cd73531e7a9cb6ed2b464767f0ee335a