Test EtcdWatcher's resilience to etcd server restart
This test reproduces a customer-reported problem where our Neutron
driver's Felix status watching was not resilient to the etcd server
being restarted.  The impact of that was that the Neutron driver
thought (after a following timeout period) that all the Felixes were
dead, and so refused to schedule any new VM instances.