Use requirements fork with pinned setuptools
setuptools 58.0.0 recently came out, and it's incompatible with some
of the Python packages - way down the food chain, and completely
unrelated AFAICT to Calico - that are pulled in as part of our
DevStack CI testing.

I think anyone using DevStack with the stable/ussuri branch will be
seeing a similar problem, so I've proposed
https://review.opendev.org/c/openstack/requirements/+/810859 to fix
the problem by pinning setuptools <58.0.0 for stable/ussuri.